modified App.jsx and changed a few things. #1 I moved header and footer directly to the Components folder and not to the nested UI folder inside of components, I deleted layout since it was an extra file and not needed.
