+++ src/App.jsx
@@ -3,8 +3,8 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 //Pages
 import HomePage from "./pages/HomePage";
-import Header from "./pages/Header";
-import Footer from "./pages/Footer";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
 
 const router = createBrowserRouter([
     {
